scraper/requirements.txt
click==8.1.2
Flask==2.1.1
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
zipp==3.8.0
certifi==2021.10.8
charset-normalizer==2.0.7
cssselect==1.1.0
idna==3.3
lxml==4.6.4
parsel==1.6.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
w3lib==1.22.0
google-cloud-storage==2.3.0
